Ez_Learn/requirements.txt
Django>=4.2,<5
gunicorn>=20
dj-database-url>=1.0
psycopg2-binary==2.9.10
whitenoise>=6.0
python-dotenv>=1.0
pytz
razorpay
xhtml2pdf
Pillow
django-crispy-forms
crispy-bootstrap5
django-bootstrap5
django-allauth
djangorestframework